HBP-151 Code review
