+++ framework_tvb/tvb/interfaces/web/controllers/simulator/simulator_controller.py
@@ -109,12 +109,6 @@ class SimulatorController(BurstBaseController):
                 self.reset_simulator_configuration()
         return result
 
-    @cherrypy.expose
-    @check_kube_user
-    def stop_operation_process(self, operation_id):
-        self.logger.info("Received a request to stop process for operation {}".format(operation_id))
-        StandAloneClient.stop_operation_process(operation_id)
-
     @expose_page
     @settings
     @context_selected
